# 🔒 CRITICAL: Stripe Payment Security Audit & Vulnerability Fixes