--- a/current-focus.md
+++ b/current-focus.md
@@ -1,42 +1,57 @@
-# Current Focus
+# Current Focus: Stripe Payment Security & Implementation Audit
 
-*Last updated: 2025-09-08 15:08:08*
+*Last updated: 2025-09-08 15:55:11*
 
 ## Session Status
 
-**Current Issue**: Update referral data in ReferralSection.tsx to reflect real reward data
-**Status**: 🔄 IN PROGRESS - Investigating RewardConfiguration and updating referral system to use accurate reward data
+**Current Issue**: Comprehensive Stripe implementation security audit and defect analysis
+**Status**: 🔍 CRITICAL SECURITY REVIEW - Analyzing payment system for vulnerabilities, defects, and compliance issues
 
 ## Context Overview
 
-from RewardConfiguration about referral reward, I wanted to update the referral data in the @src/components/referral/ReferralSection.tsx to reflect to real reward data, and wanted to make sure that the referral system is working properly
+Conducting a comprehensive security audit of the entire Stripe payment implementation to identify defects, malfunctions, security vulnerabilities, and critical issues that need immediate attention and resolution.
 
-### Current Analysis
+### Audit Scope
 
-**Component in Focus**:
-- `src/components/referral/ReferralSection.tsx` - Main referral component with hardcoded reward values
+**Core Components Under Review**:
+- `/api/payments/create-payment-intent.ts` - Payment creation endpoint
+- `/api/payments/webhook.ts` - Stripe webhook handler  
+- `/api/payments/history.ts` - Payment transaction history
+- Frontend payment components and flows
+- Database payment schemas and models
+- Environment configuration and secrets
 
-**Current Hardcoded Values**:
-- New friend gets: "+1 star + 5 coins" (line 156)  
-- User gets: "+2 stars + 5 coins" (line 162)
+**Security Focus Areas**:
+- API key exposure and rotation policies
+- Webhook signature verification integrity
+- Payment data sanitization and validation
+- Transaction integrity and atomicity
+- Error handling security implications
+- CORS configuration and authentication
+- PCI DSS compliance adherence
 
-**Investigation Needed**:
-- Find RewardConfiguration location and structure
-- Understand actual reward data format
-- Identify referral reward API endpoints
-- Verify referral system functionality
+### Investigation Strategy
 
-### Implementation Plan
+1. **Code Security Analysis**: Deep dive into all payment-related code
+2. **Flow Vulnerability Testing**: End-to-end payment lifecycle security
+3. **Data Validation Review**: Input sanitization and validation gaps
+4. **Error Scenario Testing**: Security implications of failure modes
+5. **Integration Security**: Webhook reliability and security
+6. **Compliance Audit**: PCI DSS and security standard adherence
 
-1. **Locate RewardConfiguration**: Find the source of truth for reward data
-2. **Analyze Current API**: Check `/api/referrals/status` and related endpoints  
-3. **Update Component**: Replace hardcoded values with dynamic reward data
-4. **Test Integration**: Ensure referral system works with real data
-5. **Validate Functionality**: Confirm referral rewards are calculated correctly
+## Critical Areas of Concern
 
-## Requirements
+- **Secret Management**: API key handling and exposure risks
+- **Webhook Security**: Signature verification and replay attack prevention  
+- **Transaction Integrity**: Race conditions and double-processing
+- **Data Protection**: PII handling and storage security
+- **Error Disclosure**: Information leakage in error responses
+- **Authentication**: Payment endpoint access control
 
-- **Dynamic reward display** from RewardConfiguration
-- **Accurate referral reward values**
-- **Proper integration with referral API**
-- **Maintain existing UI/UX**
+## Expected Deliverables
+
+- Security vulnerability assessment report
+- Critical issue prioritization and impact analysis
+- Implementation defect catalog with severity ratings
+- Security enhancement recommendations
+- Compliance gap analysis and remediation plan