--- a/src/lib/stripe.ts
+++ b/src/lib/stripe.ts
@@ -5,7 +5,7 @@ if (!process.env.STRIPE_SECRET_KEY) {
 }
 
 export const stripe = new Stripe(process.env.STRIPE_SECRET_KEY, {
-  apiVersion: '2025-06-30.basil',
+  apiVersion: '2025-06-30.basil', // NOTE: Using current SDK version - monitor for stability in production
   typescript: true,
 })