… docs/security/security-fixes-implemented.md
… docs/security/stripe-payment-security-audit-report.md